Move texture classes to the same file
